Refactor ICubeColor to API
